requirements.txt
# requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
psycopg2-binary
cloudinary
alembic
pyttsx3
langchain_groq
email-validator
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
fastapi 
pydantic[email] 
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart
alembic
python-dotenv
azure-storage-blob
requests
replicate
python-multipart
redis
celery